inside/requirements.txt
certifi==2018.8.13
chardet==3.0.4
click==6.7
Flask==1.0.2
flask-mongoengine==0.9.5
Flask-OAuthlib==0.9.5
Flask-WTF==0.14.2
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
mongoengine==0.15.3
oauthlib==2.1.0
pymongo==3.7.1
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
urllib3==1.23
Werkzeug==0.14.1
WTForms==2.2.1
grpcio==1.6.0